This is Code First without using Typegraph but using the native @nestjs/graphql.
