+++ app/package.json
@@ -44,6 +44,7 @@
     "@prisma/client": "^2.13.1",
     "@types/graphql-fields": "^1.3.3",
     "apollo-server-express": "^2.19.0",
+    "class-transformer": "^0.3.2",
     "class-validator": "^0.12.2",
     "graphql": "^15.5.0",
     "graphql-fields": "^2.0.3",
